pyproject.toml
[project]
name = "liine-take-home"
version = "0.1.0"
description = ""
requires-python = ">=3.10,<4"
dependencies = [
    "django (>=5.2,<6.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "django-extensions (>=4.0,<5.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pendulum (>=3.0.0,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipython = "8.35.0"
isort = "^6.0.1"

